+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - change hasSuperUser to DomainType
 - Open domain's admin is changed from zero address to normal address
+- Recon configuration
 
 ## v0.2.5